+++ src/main/java/com/security/repository/LoanRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.security.model.Loans;
 
-public interface LoanRepository  extends CrudRepository<Loans, Long>{
+public interface LoanRepository  extends CrudRepository<Loans, Integer>{
 	
 	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
